Minor fixes
Don't require a mutable Graph to determine a Heading (this appears to
have been an oversight). Move computeBearing from route.h to util.h so
users don't need to drag in Boost.
